Update setup.py using desispec template
