+++ setup.py
@@ -1,75 +1,75 @@
 #!/usr/bin/env python
-#from __future__ import absolute_import, print_function
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+from __future__ import absolute_import, division, print_function
+#
+# Standard imports
+#
 import glob
 import os
-import re
-from subprocess import Popen, PIPE
-from setuptools import setup, Command, find_packages
-
-
-def update_version_py():
-    if not os.path.isdir(".git"):
-        print("This is not a git repository.")
-        return
-    try:
-        p = Popen(["git", "describe", "--tags", "--dirty", "--always"], stdout=PIPE)
-    except EnvironmentError:
-        print("unable to run git, leaving py/desisurvey/_version.py alone")
-        return
-    out = p.communicate()[0]
-    ver = out.rstrip()
-    if p.returncode != 0:
-        print("unable to run git, leaving py/desisurvey/_version.py alone")
-        return
-    with open("py/desisurvey/_version.py", "w") as f:
-        f.write( '__version__ = \'{}\''.format( ver ) )
-    print("Set py/desisurvey/_version.py to {}".format( ver ))
-
-
-def get_version():
-    if not os.path.isfile("py/desisurvey/_version.py"):
-        print('Creating initial version file')
-        update_version_py()
-    ver = 'unknown'
-    with open("py/desisurvey/_version.py", "r") as f:
-        for line in f.readlines():
-            mo = re.match("__version__ = '(.*)'", line)
-            if mo:
-                ver = mo.group(1)
-    return ver
-
-
-class Version(Command):
-    description = "update _version.py from git repo"
-    user_options = []
-    boolean_options = []
-    def initialize_options(self):
-        pass
-    def finalize_options(self):
-        pass
-    def run(self):
-        update_version_py()
-        ver = get_version()
-        print("Version is now {}".format( ver ))
-
-
-current_version = get_version()
-
-setup (
-    name='desisurvey',
-    provides='desisurvey',
-    version=current_version,
-    description='DESI survey planning tools',
-    author='DESI Collaboration',
-    author_email='desi-data@desi.lbl.gov',
-    url='https://github.com/desihub/desisurvey',
-    package_dir={'':'py'},
-    packages=find_packages('py'),
-    scripts=[ fname for fname in glob.glob(os.path.join('bin', '*')) ],
-    license='BSD',
-    requires=['Python (>3.5.2)', ],
-    use_2to3=False,
-    zip_safe=False,
-    cmdclass={'version': Version},
-    test_suite='desisurvey.test.test_suite'
-)
+import sys
+#
+# setuptools' sdist command ignores MANIFEST.in
+#
+from distutils.command.sdist import sdist as DistutilsSdist
+from setuptools import setup, find_packages
+#
+# DESI support code.
+#
+from desiutil.setup import DesiTest, DesiVersion, get_version
+#
+# Begin setup
+#
+setup_keywords = dict()
+#
+# THESE SETTINGS NEED TO BE CHANGED FOR EVERY PRODUCT.
+#
+setup_keywords['name'] = 'desisurvey'
+setup_keywords['description'] = 'DESI Survey Planning and Scheduler'
+setup_keywords['author'] = 'DESI Collaboration'
+setup_keywords['author_email'] = 'desi-data@desi.lbl.gov'
+setup_keywords['license'] = 'BSD'
+setup_keywords['url'] = 'https://github.com/desihub/desisurvey'
+#
+# END OF SETTINGS THAT NEED TO BE CHANGED.
+#
+setup_keywords['version'] = get_version(setup_keywords['name'])
+#
+# Use README.rst as long_description.
+#
+setup_keywords['long_description'] = ''
+if os.path.exists('README.rst'):
+    with open('README.rst') as readme:
+        setup_keywords['long_description'] = readme.read()
+#
+# Set other keywords for the setup function.  These are automated, & should
+# be left alone unless you are an expert.
+#
+# Treat everything in bin/ except *.rst as a script to be installed.
+#
+if os.path.isdir('bin'):
+    setup_keywords['scripts'] = [fname for fname in glob.glob(os.path.join('bin', '*'))
+        if not os.path.basename(fname).endswith('.rst')]
+setup_keywords['provides'] = [setup_keywords['name']]
+setup_keywords['requires'] = ['Python (>2.7.0)']
+# setup_keywords['install_requires'] = ['Python (>2.7.0)']
+setup_keywords['zip_safe'] = False
+setup_keywords['use_2to3'] = False
+setup_keywords['packages'] = find_packages('py')
+setup_keywords['package_dir'] = {'':'py'}
+setup_keywords['cmdclass'] = {'version': DesiVersion, 'test': DesiTest, 'sdist': DistutilsSdist}
+setup_keywords['test_suite']='{name}.test.{name}_test_suite.{name}_test_suite'.format(**setup_keywords)
+#
+# Autogenerate command-line scripts.
+#
+# setup_keywords['entry_points'] = {'console_scripts':['desiInstall = desiutil.install.main:main']}
+#
+# Add internal data directories.
+#
+setup_keywords['package_data'] = {'desisurvey': ['data/*.yaml',
+                                                 'data/*.fits',
+                                                 'data/*.csv',
+                                                 'data/*.ecsv',]}
+#
+# Run setup command.
+#
+setup(**setup_keywords)
